Various good fixes from /analyze C++ Core Checks
After enabling the C++ Core Checks for /analyze I tried to get UIforETW
to build cleanly. This commit contains the most important issues that
were found, directly or indirectly. The changes include:

- An uninitialized member variable - harmless but risky.
- A cast in an outputPrintf call where the write type-specifier was a
better fix.
- A call to PdhCloseQuery with a nullptr
- Type punning of enums to ints without even a size check (now there is
a size check but the type punning remains for now)
- A bizarre settings bug where sizeof(LPBYTE) was used where
sizeof(uint64_t) was apparently intended, so the code only worked on
64-bit builds!
- Some overly complicated code in enableAggressiveProcessMitigations
that introduced extra pointer copies for no reason.
- ElapsedSeconds was using float math to compute a double result. There
was no warning for this (missed opportunity!) except that the C-style
cast triggered a warning on that line.

+++ UIforETW/ChildProcess.cpp
@@ -157,7 +157,7 @@ bool ChildProcess::Run(bool showCommand, std::wstring args)
 	}
 	else
 	{
-		outputPrintf(L"Error %d starting %s, %s\n", (int)GetLastError(), exePath_.c_str(), args.c_str());
+		outputPrintf(L"Error %lu starting %s, %s\n", GetLastError(), exePath_.c_str(), args.c_str());
 	}
 
 	return false;
